+++ CHANGELOG.md
@@ -8,6 +8,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 (Poikilos)
 ### Changed
 - Detect height changes from G0 commands.
-- Ensure that first temperature is not skipped
+- Ensure that first temperature is not skipped.
 - Improve GUI messages.
 - Load default path in GUI automatically.